Re-org post-processor params like Brian suggested.
